+++ journal/bounds.go
@@ -7,29 +7,29 @@ import (
 
 // IsFresh returns true if j has never contained any records.
 func IsFresh[T any](ctx context.Context, j Journal[T]) (bool, error) {
-	_, end, err := j.Bounds(ctx)
-	return end == 0, err
+	bounds, err := j.Bounds(ctx)
+	return bounds.End == 0, err
 }
 
 // IsEmpty returns true if j does not currently contain any records.
 func IsEmpty[T any](ctx context.Context, j Journal[T]) (bool, error) {
-	begin, end, err := j.Bounds(ctx)
-	return begin == end, err
+	bounds, err := j.Bounds(ctx)
+	return bounds.IsEmpty(), err
 }
 
 // FirstRecord returns the oldest record in a journal.
 func FirstRecord[T any](ctx context.Context, j Journal[T]) (Position, T, bool, error) {
 	for {
-		begin, end, err := j.Bounds(ctx)
-		if begin == end || err != nil {
+		bounds, err := j.Bounds(ctx)
+		if bounds.IsEmpty() || err != nil {
 			var zero T
 			return 0, zero, false, err
 		}
 
-		rec, err := j.Get(ctx, begin)
+		rec, err := j.Get(ctx, bounds.Begin)
 
 		if !errors.Is(err, ErrNotFound) {
-			return begin, rec, true, err
+			return bounds.Begin, rec, true, err
 		}
 
 		// We didn't find the record. Assuming the journal is not corrupted,
@@ -41,13 +41,13 @@ func FirstRecord[T any](ctx context.Context, j Journal[T]) (Position, T, bool, e
 // LastRecord returns the newest record in a journal.
 func LastRecord[T any](ctx context.Context, j Journal[T]) (Position, T, bool, error) {
 	for {
-		begin, end, err := j.Bounds(ctx)
-		if begin == end || err != nil {
+		bounds, err := j.Bounds(ctx)
+		if bounds.IsEmpty() || err != nil {
 			var zero T
 			return 0, zero, false, err
 		}
 
-		pos := end - 1
+		pos := bounds.End - 1
 		rec, err := j.Get(ctx, pos)
 
 		if !errors.Is(err, ErrNotFound) {